deep-agents-app/pyproject.toml
[project]
name = "deep-agents-app"
version = "0.1.0"
description = "Research assistant using deep-agents with subagents and Lakebase memory"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1",
    "mlflow>=3.0.0",
    "langchain-core>=0.3.0",
    "databricks-langchain[memory]>=0.16.0",
    "langgraph>=0.3.0",
    "databricks-sdk>=0.40.0",
    "databricks-mcp>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
start-app = "start_server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["start_server", "agent", "eval"]
